+++ src/elements/SendoutElement.php
@@ -829,7 +829,7 @@ class SendoutElement extends Element
         }
 
         return ContactElement::find()
-            ->select('email')
+            ->select(['email'])
             ->id($this->notificationContactIds)
             ->column();
     }